Break dependency on ij-commands
Ideally, the commands being called would be available as core service
API calls. Until then, let's call the commands by name instead of by
class, to avoid an explicit dependency.

This is admittedly very dirty since it will cause the relevant functions
to fail in the case of imagej-plugins-commands not being on the
classpath. But it is supposed to be a temporary situation...