Add Fiji's Auto-Import feature
In Fiji, many (but not all) script languages support the
"auto-import" feature, i.e. all of ImageJ 1.x classes (and for
backwards compatibility, even TrakEM2 classes) are imported by
default.

This feature is not really robust: what to do if two classes of the
same name live in two different packages? This question is not at all
academic, as the example of ImageJ 1.x' own Resizer classes show.

Another problem occurs when new classes are introduced that make
formerly unique names ambiguous. All of a sudden, all of the scripts
that use this class fail.

Basically, auto-imports share all the downsides with the highly
deprecated wildcard imports, so we will deprecate this feature right
from its introduction.

However, we need to support it so that Fiji can switch to the forked and
improved ImageJ2 script editor.

Fiji's Script Editor auto-imports all of ImageJ 1.x' classes; however,
we will *not* make imagej-ui-swing depend on ImageJ 1.x *just* to
support a deprecated feature. Instead, we introduce a new type of
plugin -- AutoImporter -- that will be implemented by a class in
imagej-legacy and enumerate the classes to auto-import.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>