TextEditor: suggest appropriate name when saving new files
When saving a Java class' source, we already have code to determine the
class name and thereby the appropriate path into which to save the file.
However, in the translation, the path got lost and the JFileChooser
got the suggestion 'null' instead.

This change fixes https://github.com/imagej/imagej-ui-swing/issues/8.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>