Updater: Replace ad-hoc popup menu by a proper one
The only reason why we tried to roll our own code to intercept the mouse
event and show our own popup menu was because the popup menu is actually
context dependent on the selected files.

However... there is a much better way to do it. We can install the popup
menu in the recommended way and then *intercept* just before it is shown
so that we can populate the menu properly.

This change incidentally fixes the problem described in
https://github.com/imagej/imagej-ui-swing/issues/2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>