When updating the updater, do not delete currently used .jar files
In the old Fiji days, we used to overwrite the .jar files when updating
the updater, but that only really worked because the file names did not
have version numbers in them. Now they do, which means that files that
are currently in use are deleted (never mind that this is not possible
on Windows).

So let's change the strategy and *not* overwrite those files, but instead
run the updater in-place from the files in update/ that will be moved
into their final destinations upon restart.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>