When updating the updater, do not loop infinitely
When the updater updated itself, it mistakenly thought that the files
were not yet downloaded and tried downloading them in an infinite loop.

The trick, of course, is to remember that we are in the middle of
updating the updater itself, and disable the check.

This issue has been reported by Claudio Collinet in
http://fiji.sc/bugzilla/show_bug.cgi?id=763 and on the mailing list.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>