Make sure that *some* files are shown in advanced mode
When the file table is empty and the user switches to advanced mode,
experience tells us that they are looking for a specific file. The
best idea is to show all the files, and already sort them by status,
so that the user only has to type in part of the file name into the
search box.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>