Script Editor: use the framework to choose files
This calls the UIService's chooseFile method, rather than hardcoding
usage of JFileChooser. This is especially important for scenarios like
the LegacyUI being visible, since in many ways it reuses the SwingUI,
but with many functions implemented differently, including chooseFile.

This slightly changes the titles of the file chooser dialog boxes,
because the SciJava UIService currently does not have a mechanism to
customize the titles. It also drops the file extension filters
previously used, since again, the UIService does not support file
filtering yet. Some day this feature can be added to SciJava Common,
but in the meantime, this change is still a net win for usability.