Let's call it the ImageJ Wiki in the ImageJ updater
It is really funny how some people still think that Fiji is something
completely different than ImageJ. Granted, the name is different, but
we were asked specifically not to use the same name. Oh well, history.

Anyway, the ImageJ Wiki -- http://wiki.imagej.net/ -- is just another
skin and default page for the Fiji Wiki, to reflect the nature of the
true relationship between the Fiji and the ImageJ2 project: very,
very collaborative and collegial.

So let's just tell the ImageJ updater users that they are about to
get an ImageJ Wiki account for the to-be-initialized personal update
site.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>