Update Sites: do not handle key events from just-closed dialog
We should give the user a chance to provide a password before we try
to log into the Wiki. But when the user hit "Enter" to dismiss the
message that a Wiki account was registered and needs to be activated,
the "key released" event happens already after the message dialog is
dismissed, at least on MacOSX. To avoid misinterpreting this event as
if it originated from the password field, let's handle the "key
pressed" event instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>