InterpreterWindow: make each language a tab
This is much simpler than trying to keep track of an "active" language,
shuffling things around, etc. In this way, all the available languages
are accessible from a single click. It is unlikely the number of
available scripting languages will grow too large to accommodate this
style of UI.

This also keeps the UI simple, eliminating the need for MigLayout.