SitesDialog: fix appearing of previous URL
Removing a row that was still edited from the manage update sites dialog
resulted in validity checks which are actually unnecessary for a site
that is removed and prohibited proper stopping of the editing process.
This in turn resulted in the URL of a previous editing to reapper once
you add a new update site.