ImageJUpdater: Do not update the updater separately
The removed code does the following:
  * check if the updater can be updated
  * if yes, download updater and it's dependencies and install it,
    do not ask for approval
  * try to restart the updater in place
  * if that does not work, ask for restart of Fiji

This is why it should be removed:
  * if other components depend on the updater, this can break the installation
    because the other component will not get updated at the same time as the updater
  * the updater will be updated without the users approval

CTR: I asked Johannes why this logic was introduced. Here is what he said:

> The reason was dependency hell. If you don't update the updater first,
> then create a new class loader, then re-load the updater and run it,
> it really can break everything, especially when imagej-ui-swing is
> updated (and not necessarily the updater).

CTR: So really, there are things that can go wrong either way. In the
interest of simplicity, we remove the logic. If all hell breaks loose,
we can put it back, and deal with the fallout as best we can. But I
expect that removing this logic will fix more problems than it causes.

Closes #83.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>