UpdaterFrame: remove eager git executable check
Unfortunately, on macOS, trying to invoke git (which will be
/usr/bin/git on a fresh system) triggers Apple's wonderful "you ran
a command line tool for something that isn't installed so we're going to
pop a graphical dialog in response" genius tactic, made most famous in
our community previously by doing that in response to attempting to
breathe on Apple Java's shared library without it being installed.

Let's not make it so that every single Mac user is prompted to install
the Xcode developer command line tools just to update their ImageJ2.

"But Curtis!" you might ask. "What happens if git is not available on a
particular system, and you are no longer checking for it up front? Does
everything crash and burn?" It turns out everything is OK.

What happens is: if you click "Show changes" to pop the diff dialog,
then click the "Hex Diff" or "Class File Diff" button, the output pane
will show many errors of the form:

  /Java-8/jars/imagej-updater-0.10.7.jar-20220314155634
  java.lang.RuntimeException: java.io.IOException: Cannot run program "git": error=2, No such file or directory
	  at org.scijava.util.ProcessUtils.exec(ProcessUtils.java:135)
	  at org.scijava.util.ProcessUtils.exec(ProcessUtils.java:69)
	  at net.imagej.updater.Diff.showDiff(Diff.java:273)
	  at net.imagej.updater.Diff.showDiff(Diff.java:195)
	  at net.imagej.ui.swing.updater.DiffFile$2.run(DiffFile.java:134)
	  at net.imagej.ui.swing.updater.DiffFile$3.run(DiffFile.java:160)

which IMHO is good enough to clue in savvy people that they need git.
Because most people won't click on buttons with those labels.

If someone complains, we can make the error handling more pleasing then.