Add mechanism for updating Java
We can now control the java version per-platform using
https://downloads.imagej.net/java/jdk-urls.txt, which contains URLs for
JDK downloads. These are cached locally and any time a more-recent
version is available, it is downloaded to the appropriate /java
sub-directory.

If jaunch is being used, an app.cfg file is created to tell the launcher
which JDK to use on subsequent launches.