Put restart failure logs in a subdir
Also timestamps them instead of checking if the file exists and only
writing one.

I'd prefer to simply not create empty error log files but the timing of
then needing to keep the ProcessBuilder's (sub)process alive until the
script (or really just startup) has finished to then check the length of
the file is overly complicated and feels extremely fragile. So I would
rather just hide the files in a subdir so it isn't painfully obvious to
anyone glancing at the Fiji folder.

Using timestamps instead of checking for the existing files also seemed
reasonable since we're a) not filling up the base directory and b) these
files are only created during the upgrade process which is unlikely to
take place multiple times (and if it does then all those logs could be
valuable)