Use file-based output to check launcher
On Windows, if we launch with Jaunch it is a GUI-based app which means
the system console is null and any subprocess will silently eat its
output. Thus we need to use a temp file to store and read the Props
output.

NB: this requres the updated Props.class in
https://github.com/apposed/jaunch/commit/832dc164e18d9121712f4942b0743e3c44040b81