Force arm64 native even if using rosetta
Double-clicking the Fiji.app launcher will use the universal mac
launcher, which will select arm64 on arm64 macs. There is no way to
change this or allow the user to opt into arm64 native due to the atomic
nature of the signed Fiji.app bundle. Thus, if you're on arm64 and
upgrade, you get arm64 native.